fix(compat): clean SDL_QUIT shutdown, volatile flag, int32_t timers, copybufreverse underrun
compat/sdl_driver.{c,h}: replace direct exit(0) on SDL_QUIT with a clean
shutdown path. SDL_QUIT now sets a module-static
'volatile sig_atomic_t sdl_quit_requested' flag and the existing engine
sdl_checkquit() poll picks it up to break out of the main loop at
source/GAME.C, allowing normal cleanup. Add sdl_quit_requested_get() to
sdl_driver.h.

compat/sdl_driver.{c,h} + compat/mact_stub.c + compat/pragmas_gcc.h:
convert sdl_getbytesperline(), sdl_getticks(), gettime1mhz(),
deltatime1mhz(), readtimer() from 'long' (platform-dependent) to int32_t
fixed-width. Adds _Static_assert(sizeof(int32_t) == 4) for compile-time
ABI safety.

compat/pragmas_gcc.h: fix copybufreverse buffer underrun — the reverse-
copy was indexing 'src[-i]' (reads before start of buffer) when it should
be 'src[n-1-i]'. Function is currently unreachable in active code but the
fix is defense-in-depth.

+++ compat/mact_stub.c
@@ -360,9 +360,9 @@ long MOUSE_GetButtons(void) {
 /* Timer stubs if not already defined */
 void inittimer1mhz(void) { }
 void uninittimer1mhz(void) { }
-long gettime1mhz(void) { return sdl_getticks() * 1000; }
-long deltatime1mhz(void) { return 0; }
-long readtimer(void) { return sdl_getticks(); }
+int32_t gettime1mhz(void) { return sdl_getticks() * 1000; }
+int32_t deltatime1mhz(void) { return 0; }
+int32_t readtimer(void) { return sdl_getticks(); }
 
 /* Callback stub */
 void testcallback(unsigned long val) { (void)val; }
